fix stuff to make clang happy

+++ fheco/src/trs/matching_term.cpp
@@ -8,11 +8,13 @@ namespace fheco_trs
 size_t MatchingTerm::term_id = 0;
 
 MatchingTerm::MatchingTerm(int64_t _value) : term_type(TermType::scalarType), value(_value), id(term_id++) {}
-MatchingTerm::MatchingTerm(int _value) : term_type(TermType::scalarType), value(_value), id(term_id++) {}
+MatchingTerm::MatchingTerm(int _value)
+  : term_type(TermType::scalarType), value(static_cast<int64_t>(_value)), id(term_id++)
+{}
 MatchingTerm::MatchingTerm(double _value) : term_type(fheco_trs::TermType::scalarType), value(_value), id(term_id++) {}
 
 MatchingTerm::MatchingTerm(bool _value)
-  : term_type(TermType::booleanType), value(static_cast<int>(_value)), id(term_id++)
+  : term_type(TermType::booleanType), value(static_cast<int64_t>(_value)), id(term_id++)
 {}
 
 MatchingTerm::MatchingTerm(FunctionId func_id) : term_type(fheco_trs::TermType::functionType), function_id(func_id) {}
